Update lacrosse task to family layout
