+++ tasks/agentic_vbench_understanding/lacrosse-mich-jhu-2024/calibration/run_claude_calibration.sh
