+++ tasks/agentic_vbench_understanding/lacrosse-mich-jhu-2024/calibration/runpack/audit_and_grade.sh
