+++ tasks/agentic_vbench_understanding/lacrosse-mich-jhu-2024/calibration/runpack/run_gemini.sh
