+++ tasks/agentic_vbench_understanding/lacrosse-mich-jhu-2024/steps/solve/solution/solve.sh
