+++ tasks/agentic_vbench_understanding/lacrosse-mich-jhu-2024/steps/solve/tests/test.sh
@@ -1,0 +1,14 @@
+#!/bin/bash
+# Verifier entry point. Preserve the agent's output for the audit trail, then grade.
+set -euo pipefail
+
+mkdir -p /logs/verifier /logs/artifacts
+
+if [ -d /workspace/output ]; then
+    cp -a /workspace/output/. /logs/artifacts/ 2>/dev/null || true
+fi
+
+python3 /tests/judge.py \
+        --solution /workspace/output/solution.json \
+        --reward-json /logs/verifier/reward.json \
+        --reward-txt /logs/verifier/reward.txt
