+++ tasks/agentic_vbench_understanding/lacrosse-mich-jhu-2024/task.toml
@@ -1,0 +1,40 @@
+version = "1.0"
+
+[task]
+name = "agentic-vbench/mich-jhu-2024-lacrosse-goal-ledger"
+
+[metadata]
+difficulty = "hard"
+category = "video-understanding"
+tags = ["video-understanding", "sports-analytics", "lacrosse", "event-timeline", "agentic-vbench-sports"]
+# Full broadcast of one NCAA D-I men's lacrosse game, masked and muted:
+# the entire lower third is blacked out (no scoreboard, no score cards, no
+# player-bio graphics) and the audio + container metadata are stripped.
+# Ground truth: the game's official NCAA box score / play-by-play, with jersey
+# numbers mapped via both teams' official rosters, then audited frame-by-frame
+# against this exact encode (26/26 goals located in key order).
+source = "public full-game broadcast upload, re-encoded with the lower third masked; SHA-256 pinned"
+
+[environment]
+build_timeout_sec = 2400.0
+cpus = 4
+memory_mb = 8192
+# ~1 GB masked broadcast, plus room for the frames the agent samples.
+storage_mb = 8192
+# The clip is baked at build time and the goal-by-goal ledger is not in any
+# public record at the scored granularity: keep this false.
+allow_internet = false
+
+[environment.env]
+HF_TOKEN = "${HF_TOKEN:-}"
+
+[[steps]]
+name = "solve"
+
+[steps.agent]
+# A full 102-minute game: measured genuine attempts ran 95+ minutes and
+# 100-1200+ tool-call turns, so give the agent real time to seek and tally.
+timeout_sec = 5400.0
+
+[steps.verifier]
+timeout_sec = 600.0
